requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
PyMuPDF==1.23.8
Pillow==10.1.0
pytesseract==0.3.10
python-multipart==0.0.6
